+++ lib/views/view_loans/view_loans.dart
@@ -44,8 +44,11 @@ class ViewLoansState extends ViewWidgetState<LoanPayment> {
   }
 
   @override
-  Widget getPanelForTransactions(final List<int> indices) {
-    return _getSubViewContentForTransactions(indices);
+  Widget getPanelForTransactions({
+    required final List<int> selectedItems,
+    required final bool showAsNativeCurrency,
+  }) {
+    return _getSubViewContentForTransactions(selectedItems);
   }
 
   @override