+++ lib/views/view_payees/view_payees.dart
@@ -46,8 +46,11 @@ class ViewPayeesState extends ViewWidgetState<Payee> {
   }
 
   @override
-  Widget getPanelForTransactions(final List<int> indices) {
-    return _getSubViewContentForTransactions(indices);
+  Widget getPanelForTransactions({
+    required final List<int> selectedItems,
+    required final bool showAsNativeCurrency,
+  }) {
+    return _getSubViewContentForTransactions(selectedItems);
   }
 
   @override