+++ lib/views/view_rentals/view_rentals.dart
@@ -57,8 +57,11 @@ class ViewRentalsState extends ViewWidgetState<RentBuilding> {
   }
 
   @override
-  Widget getPanelForTransactions(final List<int> indices) {
-    return _getSubViewContentForTransactions(indices);
+  Widget getPanelForTransactions({
+    required final List<int> selectedItems,
+    required final bool showAsNativeCurrency,
+  }) {
+    return _getSubViewContentForTransactions(selectedItems);
   }
 
   @override